Updated binding.
The web server running in your container is listening for connections on port 5000 of the loopback network interface (127.0.0.1). As such this web server will only respond to http requests originating from that container itself.
In order for the web server to accept connections originating from outside of the container you need to have it bind to the 0.0.0.0 IP address.